+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Bug fixes:
 * Check that `Symbol`s are always created from a non-broken `String` (@eregon).
 * Fix converting the same mutable `String` to native memory concurrently (@eregon).
 * Fix keeping `Float` and big `Integer` elements of an `Array` converted to native storage by `RARRAY_PTR()` alive (@eregon).
+* Fix `Marshal.load` for an object extended with a module containing a nested user-marshaled object (#3943, @andrykonchin).
 
 Compatibility: